fix(frontend): move NAV_LINKS para arquivo separado do Header
O ESLint (react-refresh/only-export-components) barra arquivos de
componente que exportam algo além de componentes React, pois isso
quebra o Fast Refresh. Header.tsx exportava a constante NAV_LINKS
junto do componente, o que falhava o Lint & Build no CI.
