+++ frontend/src/components/Home/nav-links.ts
@@ -1,0 +1,7 @@
+export const NAV_LINKS = [
+  { label: "Início", to: "/" },
+  { label: "Dashboard", to: "/dashboard" },
+  { label: "Editor", to: null },
+  { label: "Vagas", to: null },
+  { label: "Preços", to: null },
+];
